FFI Native Method Handler still contains mangled entrypoint, but if I remove it, it works

+++ ffi-test-csharp/src/deltalake.kernel.rust.interop/DeltaLake/Kernel/Rust/Ffi/FileMeta.cs
@@ -1,3 +1,10 @@
+// -----------------------------------------------------------------------------
+// <copyright company="The Delta Lake Project Authors">
+// Copyright (2024) The Delta Lake Project Authors.  All rights reserved.
+// Licensed under the Apache license. See LICENSE file in the project root for full license information.
+// </copyright>
+// -----------------------------------------------------------------------------
+
 namespace DeltaLake.Kernel.Rust.Ffi;
 
 public partial struct FileMeta
@@ -9,5 +16,5 @@ public partial struct FileMeta
     public long last_modified;
 
     [NativeTypeName("uintptr_t")]
-    public nuint size;
+    public ulong size;
 }
