+++ ffi-test-csharp/src/deltalake.kernel.rust.interop/DeltaLake/Kernel/Rust/Ffi/KernelStringSlice.cs
@@ -1,3 +1,10 @@
+// -----------------------------------------------------------------------------
+// <copyright company="The Delta Lake Project Authors">
+// Copyright (2024) The Delta Lake Project Authors.  All rights reserved.
+// Licensed under the Apache license. See LICENSE file in the project root for full license information.
+// </copyright>
+// -----------------------------------------------------------------------------
+
 namespace DeltaLake.Kernel.Rust.Ffi;
 
 public unsafe partial struct KernelStringSlice
@@ -6,5 +13,5 @@ public unsafe partial struct KernelStringSlice
     public sbyte* ptr;
 
     [NativeTypeName("uintptr_t")]
-    public nuint len;
+    public ulong len;
 }
